Refactor code to obey PEP8 guidelines Create the tests folder and move different tests to different files to improve on modularity
