+++ tests/test_user_model.py
@@ -1,0 +1,28 @@
+import unittest
+
+from api.models import User
+
+
+class UserModelTestCase(unittest.TestCase):
+    def setUp(self):
+        self.user = User("dennisjjagwe@gmail.com",
+                         "jjagwe", "dennis", "mypassword")
+
+    def test_user_has_no_id_before_addition_to_database(self):
+        self.assertFalse(self.user.id is not None)
+        self.assertTrue(self.user.id is None)
+
+    def test_verify_password(self):
+        self.assertTrue(self.user.verify_password("mypassword"))
+        self.assertFalse(self.user.verify_password("otherpassword"))
+
+    def test_set_password(self):
+        self.user.set_password("new_password")
+        self.assertTrue(self.user.verify_password("new_password"))
+        self.assertFalse(self.user.verify_password("mypassword"))
+
+    def test_stored_passwords_are_random_if_password_strings_are_similar(self):
+        user1 = User("dennisj@gmail.com", "jjagwe", "denni", "password")
+        user2 = User("dennisj@gmail.com", "king", "dennis", "password")
+        self.assertTrue(user1.password != user2.password)
+        self.assertFalse(user2.password == user1.password)
